requirements.txt
# Naver Blog AI Detector - ML Training Requirements
# Python 3.9+
# Last Updated: 2025-10-17

# ========================================
# Core Deep Learning Framework
# ========================================
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Note: Install PyTorch with CUDA support if available:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For Apple Silicon (M1/M2/M3/M4): MPS acceleration enabled by default

# ========================================
# Transformers & NLP
# ========================================
transformers>=4.50.0
tokenizers>=0.20.0
sentencepiece>=0.2.0
datasets>=3.0.0
accelerate>=1.0.0

# Korean NLP (Optional)
# konlpy>=0.6.0  # Uncomment if using KoNLPy for Korean tokenization

# ========================================
# LLM API for Synthetic Data Generation
# ========================================
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0

# ========================================
# CoreML Conversion (macOS 14+, iOS 17+)
# ========================================
coremltools>=8.0.0  # INT4 quantization, W8A8 support

# ========================================
# ONNX Export & Optimization
# ========================================
onnx>=1.17.0
onnxruntime>=1.19.0
optimum>=1.23.0  # Hugging Face optimization toolkit

# ========================================
# Data Processing & Analysis
# ========================================
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0

# ========================================
# Visualization & Monitoring
# ========================================
matplotlib>=3.9.0
seaborn>=0.13.0
tensorboard>=2.18.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.18.0

# ========================================
# Progress Bars & Utilities
# ========================================
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
joblib>=1.3.0

# ========================================
# Web Scraping (for Data Collection)
# ========================================
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Optional: Selenium for dynamic content
# selenium>=4.15.0
# webdriver-manager>=4.0.0

# ========================================
# Testing & Code Quality
# ========================================
pytest>=8.3.0
pytest-cov>=6.0.0
black>=24.10.0
flake8>=7.1.0
isort>=5.13.0

# ========================================
# Type Hints
# ========================================
mypy>=1.13.0
types-PyYAML>=6.0.12
types-tqdm>=4.66.0
types-requests>=2.31.0

# ========================================
# System Monitoring (Optional)
# ========================================
psutil>=6.1.0
# gputil>=1.4.0  # For GPU monitoring

# ========================================
# Development Tools (Optional)
# ========================================
# jupyter>=1.0.0
# ipykernel>=6.25.0
# notebook>=7.0.0

# ========================================
# Installation Instructions
# ========================================
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # Windows: venv\Scripts\activate
#
# 2. Install dependencies:
#    pip install -r requirements.txt
#
# 3. For GPU support (NVIDIA):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# 4. Verify installation:
#    python -c "import torch; print(f'PyTorch: {torch.__version__}')"
#    python -c "import transformers; print(f'Transformers: {transformers.__version__}')"
#    python -c "import coremltools; print(f'CoreML Tools: {coremltools.__version__}')"
